requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pathway[xpack-llm]>=0.13.0,<1.0.0
docling>=2.0.0,<3.0.0
pypdf>=4.0.0,<5.0.0
pdf2image>=1.17.0,<2.0.0
unstructured[pdf]>=0.13.0,<1.0.0
sentence-transformers>=3.0.0,<4.0.0
torch>=2.2.0,<3.0.0
groq>=0.9.0,<1.0.0
requests>=2.32.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
